backend/requirements.txt
# real estate scraping and api backend

fastapi==0.104.*
uvicorn==0.24.*
sqlalchemy==2.0.*
psycopg2-binary==2.9.*
alembic==1.12.*
selenium==4.15.*
undetected-chromedriver==3.5.*
beautifulsoup4==4.12.*
requests==2.31.*
pandas==2.1.*
python-dotenv==1.0.*
pydantic==2.5.*
python-multipart==0.0.*
unidecode==1.3.*
